+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Changelog
 
-All Notable changes to `studiobonito/silverstripe-typekit` will be documented in this file
+All Notable changes to `studiobonito/silverstripe-typekit` will be documented in this file.
 
 ## [Unreleased]
+
+## [2.0.0] - 2015-12-17
+### Added
+- Created unit tests.
+- Updated module to work with SilverStripe CMS 3.2.
+- Included `.editorconfig` file.
+
+### Fixed
+- Changed namespace from `TemplateGlobalProvider` to `TemplateGlobalProviders` to match directory name.
+- Changed static call on `SiteConfig` to use `Injector` instead for better testing.
+- Increased the minimum version for `silverstripe-framework` and `silverstripe-cms` to `3.1.2`.
 
 ## [1.1.0] - 2015-10-08
 ### Added
@@ -21,6 +32,7 @@ All Notable changes to `studiobonito/silverstripe-typekit` will be documented in
 ### Added
 - Updated dependencies to include `silverstripe/framework` version `~3.1.0`.
 
-[Unreleased]: https://github.com/studiobonito/silverstripe-typekit/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/studiobonito/silverstripe-typekit/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/studiobonito/silverstripe-typekit/compare/1.1.0...2.0.0
 [1.1.0]: https://github.com/studiobonito/silverstripe-typekit/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/studiobonito/silverstripe-typekit/compare/1.0.0...1.0.1